SanskarKambleProject/requirements.txt
google-generativeai
pandas
numpy
seaborn
streamlit
matplotlib
scikit-learn